Refactor logging admin to use driver abstraction
