+++ plugins/logging/pkg/backend/sync.go
@@ -12,7 +12,7 @@ import (
 	"github.com/rancher/opni/pkg/capabilities/wellknown"
 	"github.com/rancher/opni/plugins/logging/pkg/apis/node"
 	loggingerrors "github.com/rancher/opni/plugins/logging/pkg/errors"
-	"github.com/rancher/opni/plugins/logging/pkg/gateway/drivers"
+	driver "github.com/rancher/opni/plugins/logging/pkg/gateway/drivers/backend"
 	"go.uber.org/zap"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
@@ -100,11 +100,11 @@ func (b *LoggingBackend) buildResponse(old, new *node.LoggingCapabilityConfig) *
 func (b *LoggingBackend) shouldDisableNode(ctx context.Context) bool {
 	installState := b.ClusterDriver.GetInstallStatus(ctx)
 	switch installState {
-	case drivers.Absent:
+	case driver.Absent:
 		return true
-	case drivers.Pending, drivers.Installed:
+	case driver.Pending, driver.Installed:
 		return false
-	case drivers.Error:
+	case driver.Error:
 		fallthrough
 	default:
 		// if status is unknown don't uninstall from the node
